# swagger.yml
openapi: 3.0.0
paths:
...
  /api/inventory/stock-balances:
    get:
      operationId: InventoryController_listStockBalances
      parameters:
        - name: page
          required: false
          in: query
          schema:
            $ref: '#/components/schemas/Object'
        - name: limit
          required: false
          in: query
          schema:
            $ref: '#/components/schemas/Object'
        - name: itemId
          required: false
          in: query
          schema:
            type: string
        - name: locationId
          required: false
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      security: &ref_13
        - bearer: []
      summary: List stock balances with pagination and filters
      tags: &ref_14
        - inventory
  /api/inventory/stock-movements:
    get:
      operationId: InventoryController_listStockMovements
      parameters:
        - name: page
          required: false
          in: query
          schema:
            $ref: '#/components/schemas/Object'
        - name: limit
          required: false
          in: query
          schema:
            $ref: '#/components/schemas/Object'
        - name: movementType
          required: false
          in: query
          schema:
            type: string
            enum:
              - in
              - out
              - transfer
              - adjustment
      responses:
        '200':
          description: ''
      security: *ref_13
      summary: List stock movements with pagination and optional movement type filter
      tags: *ref_14
  /api/inventory/adjustments:
    post:
      operationId: InventoryController_createAdjustment
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStockAdjustmentDto'
      responses:
        '201':
          description: ''
      security: *ref_13
      summary: Apply stock adjustment to an existing stock balance record
      tags: *ref_14
  /api/inventory/adjust-quantity:
    post:
      operationId: InventoryController_adjustQuantity
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdjustStockByReferenceDto'
      responses:
        '201':
          description: ''
      security: *ref_13
      summary: >-
        Adjust stock quantity by product and location, creating balance if
        needed
      tags: *ref_14
...
info:
  title: RxSoft Backend API
  description: Phase 3 identity module and API contracts
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    Object:
      type: object
      properties: {}
...
    CreateStockAdjustmentDto:
      type: object
      properties:
        stockBalanceId:
          type: string
        deltaQuantity:
          type: number
          description: Can be positive or negative
        reason:
          type: string
      required:
        - stockBalanceId
        - deltaQuantity
        - reason
    AdjustStockByReferenceDto:
      type: object
      properties:
        itemId:
          type: string
        locationId:
          type: string
        lotId:
          type: string
        deltaQuantity:
          type: number
        reason:
          type: string
        reorderMinQty:
          type: number
        reorderMaxQty:
          type: number
      required:
        - itemId
        - locationId
        - deltaQuantity
        - reason
...